crop ratio

+++ llava/train/train.py
@@ -74,7 +74,7 @@ class DataArguments:
     is_multimodal: bool = False
     image_folder: Optional[str] = field(default=None)
     image_aspect_ratio: str = 'square'
-    crop_ratio: int = 0.65
+    crop_ratio: float = 0.65
     crop: bool = True
